requirements.txt
pre-commit
python-dotenv
instructor==0.4.8
pdfservices-sdk==2.3.0
openai==1.13.3
instructor==0.4.8
tiktoken==0.5.2
dvc==3.6.0
dvc-s3==2.23.0
loguru==0.7.2
awscli==1.32.34
unicodedata2==15.1.0
numpy==1.26.4
pandas==2.2.0
pytz==2024.1
langchain==0.1.11
langchain-community==0.0.27
langchain-core==0.1.30
langchain-text-splitter==0.0.1
langchain_benchmarks==0.0.10
langsmith==0.1.23
langchain-openai==0.0.7
tqdm==4.66.1
rapidfuzz==3.6.2
